% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\docType{package}
\name{DROP-package}
\alias{DROP}
\alias{DROP-package}
\title{DROP: Distributionally Robust Optimization for Precision Matrices}
\description{
Implements adaptive distributionally robust optimization (DRO) for estimating high-dimensional Gaussian graphical models. The method achieves robust graph recovery through a rank-based inverse normal transformation and efficiently solves the adaptive-weighted objective function using a coordinate descent algorithm with C++ acceleration and parallel computing for improved performance.
}
\author{
\strong{Maintainer}: Canruo Shen \email{canruo.shen@ubc.ca}

}
\keyword{internal}
