% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_empty.list}
\alias{drop_empty.list}
\title{Drop list elements with empty value}
\usage{
\method{drop_empty}{list}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Drop list elements with empty value
}
