% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_empty}
\alias{drop_empty}
\title{Drop list elements with empty value}
\usage{
drop_empty(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Drop list elements with empty value
}
\details{
See:
\itemize{
\item \code{\link[=drop_empty.list]{drop_empty.list()}}
}
}
\examples{
x <- list(
    foo = list(bar = integer(), baz = list(x = 1, y = character(), z = 3)),
    bar = 2,
    baz = logical()
)
drop_empty(x)
}
