% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_null.list}
\alias{drop_null.list}
\title{Drop list elements with value \code{NULL}}
\usage{
\method{drop_null}{list}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Drop list elements with value \code{NULL}
}
