% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.R
\name{drop_null}
\alias{drop_null}
\title{Drop list elements with value \code{NULL}}
\usage{
drop_null(x)
}
\arguments{
\item{x}{}
}
\value{
(\link{list}) Cleaned list
}
\description{
Drop list elements with value \code{NULL}
}
\details{
See:
\itemize{
\item \code{\link[=drop_null.list]{drop_null.list()}}
}
}
\examples{
x <- list(
    foo = list(bar = NULL, baz = list(x = 1, y = NULL, z = 3)),
    bar = 2,
    baz = NULL
)
drop_null(x)
}
