% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-package.R
\docType{package}
\name{drop-package}
\alias{drop}
\alias{drop-package}
\title{drop: Drop stuff}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Janko Thyson \email{janko@rappster.io}

}
\keyword{internal}
