% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-package.R
\docType{package}
\name{drop-package}
\alias{drop}
\alias{drop-package}
\title{drop: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
